+++ packages/graph-explorer/src/components/radix/index.ts
@@ -1,5 +1,0 @@
-export * from "./Checkbox";
-export * from "./Form";
-export * from "./Input";
-export * from "./Label";
-export * from "./Select";